+++ mobile/android/android-components/docs/changelog.md
@@ -5,6 +5,9 @@ permalink: /changelog/
 ---
 
 # 137.0 (In Development)
+* **support-ktx**
+  * 🌟 Added `String.isContentUrl` method that checks if the string is a content URL. [Bug 1944084](https://bugzilla.mozilla.org/show_bug.cgi?id=1944084).
+  * 🌟 Added `String.isAboutUrl` method that checks if the string is an about URL. [Bug 1944084](https://bugzilla.mozilla.org/show_bug.cgi?id=1944084).
 
 # 136.0
 * **support-ktx**